Reduce dependencies to pytorch_lightning
- separate model factory using PyTorch model API
- MNIST preprocessor doesn't require pl_bolts import
