+++ perceiver/data/__init__.py
@@ -1,3 +1,2 @@
-from perceiver.data.imdb import ImdbDataModule
-from perceiver.data.mnist import MnistDataModule, MnistPreprocessor
-from perceiver.data.utils import TextPreprocessor
+from perceiver.data.mnist_preproc import MnistPreprocessor
+from perceiver.data.text_preproc import TextPreprocessor
